requirements.txt
django==4.0.4
django-sitetree==1.17.2
djangorestframework==3.13.1
django-filter==21.1
django-cors-headers==3.12.0
markdown==3.3.7
#django-downloadview==2.3.0
redis==4.3.1
hiredis==2.0.0
dramatiq==1.13.0
psycopg2-binary==2.9.3
